format the code
