+++ utility/helper.py
@@ -2,6 +2,7 @@ import os
 import json
 from pathlib import Path
 from warnings import warn
+import pygame
 
 class Helper:
     """
@@ -121,5 +122,16 @@ class Helper:
             for key, value in world_stats.items():
                 f.write(f"{key}: {value}\n")
                 print(f"{key}: {value}")
+
+    @staticmethod
+    def check_pause(env):
+        """Check if GUI is paused and wait if needed."""
+        if env.train_gui and env.gui and hasattr(env.gui, 'paused'):
+            while env.gui.paused and not env.gui.step:
+                env.gui.render(env)
+                # Break if window closed
+                if not pygame.display.get_init():
+                    return False
+        return True
